Refactored the code to add: - Sub-Coordinator concept - Move Data into a more self-contained construct - Separate screens out into their own files
