+++ Coordinator/Coordinator.swift
@@ -11,9 +11,6 @@ import SwiftUI
  
 struct MainCoordinator: View {
     
-    @EnvironmentObject private var data: Data
-    static let timer = Timer.publish(every: 1.5, on: .main, in: .common).autoconnect()
-
     enum Screen: Int {
         case first
         case second
@@ -31,12 +28,9 @@ struct MainCoordinator: View {
                 Screen2(done: self.screen2Done)
                     .transition(.forwardScreenTransition)
             } else if self.screen.rawValue == Screen.third.rawValue {
-                Screen3(items: $data.food, done: self.screen3Done)
-                    .transition(.forwardScreenTransition)
+                // Return a sub-coordinator instead of a screen
+                Screen3Coordinator(done: self.screen3Done)
             }
-        }.onReceive(MainCoordinator.timer) { _ in
-            let numberOfItems = Int.random(in: 1..<10)
-            self.data.food = Array(repeating: "🥝 Yummy!", count: numberOfItems)
         }
     }
         
@@ -58,85 +52,7 @@ struct MainCoordinator: View {
         }
     }
 }
-  
-struct Screen1: View {
-    var done: () -> Void
-    
-    var body: some View {
-        VStack(spacing: 20) {
-            Text("Home Screen")
-                .font(.largeTitle)
-            
-            Spacer()
-            
-            Button(action: done) {
-                VStack {
-                    Text("✨Go✨")
-                        .multilineTextAlignment(.center)
-                    Image(systemName: "tornado")
-                }.font(.largeTitle)
-            }
-            
-            Spacer()
-        }
-    }
-}
  
-struct Screen2: View {
-    
-    var done: () -> Void
-    
-    var body: some View {
-        VStack {
-            Text("Screen2").font(.title)
-            Button(action: {
-                self.done()
-            }) {
-                Text("Go to the tabs!")
-            }
-        }
-    }
-}
- 
-struct Screen3: View {
-    @Binding var items: [String]
-
-    var done: () -> Void
-    
-    var body: some View {
-        TabView {
-            VStack {
-                Text("Tab 1")
-                    .font(.largeTitle)
-                Button(action: done) {
-                    Text("Logout")
-                }
-            }
-            .tabItem {
-                Image(systemName: "1.square.fill")
-                Text("First")
-            }
- 
-            Text("Tab 2")
-                .font(.largeTitle)
-                .tabItem {
-                    Image(systemName: "2.square.fill")
-                    Text("Second")
-            }
- 
-            VStack {
-                Text("Tab 3")
-                    .font(.largeTitle)
-                MyTable(items: $items)
-            }
-            .tabItem {
-                Image(systemName: "3.square.fill")
-                Text("Third")
-            }
-        }
-    }
-}
-
 
 struct MainCoordinator_Previews: PreviewProvider {
     static var previews: some View {
@@ -160,14 +76,3 @@ extension AnyTransition {
     }
 }
 
-struct MyTable: View {
-    @Binding var items: [String]
-    
-    var body: some View {
-        List {
-            ForEach(items, id: \.self) { item in
-                Text("\(item)")
-            }
-        }
-    }
-}
